+++ src/Sentry/Hint.cs
@@ -6,14 +6,20 @@ namespace Sentry;
 /// </summary>
 public class Hint
 {
+    private readonly SentryOptions? _options;
     private readonly List<Attachment> _attachments = new();
     private readonly Dictionary<string, object?> _items = new();
 
     /// <summary>
     /// Creates a new instance of <see cref="Hint"/>.
     /// </summary>
-    public Hint()
+    public Hint() : this(SentrySdk.CurrentHub.GetSentryOptions())
     {
+    }
+
+    internal Hint(SentryOptions? options)
+    {
+        _options = options;
     }
 
     /// <summary>
@@ -26,13 +32,6 @@ public class Hint
     {
         _items[key] = value;
     }
-
-    /// <summary>
-    /// The Java SDK has some logic so that certain Hint types do not copy attachments from the Scope.
-    /// This provides a location that allows us to do the same in the .NET SDK in the future.
-    /// </summary>
-    /// <param name="scope">The <see cref="Scope"/> that the attachments should be copied from</param>
-    internal void AddAttachmentsFromScope(Scope scope) => _attachments.AddRange(scope.Attachments);
 
     /// <summary>
     /// Attachments added to the Hint.
@@ -51,6 +50,35 @@ public class Hint
     /// BeforeSend and others.
     /// </remarks>
     public IDictionary<string, object?> Items => _items;
+
+    /// <summary>
+    /// The Java SDK has some logic so that certain Hint types do not copy attachments from the Scope.
+    /// This provides a location that allows us to do the same in the .NET SDK in the future.
+    /// </summary>
+    /// <param name="scope">The <see cref="Scope"/> that the attachments should be copied from</param>
+    internal void AddAttachmentsFromScope(Scope scope) => _attachments.AddRange(scope.Attachments);
+
+    /// <summary>
+    /// Creates a new Hint with one or more attachments.
+    /// </summary>
+    /// <param name="filePath">The path to the file to attach.</param>
+    /// <param name="type">The type of attachment.</param>
+    /// <param name="contentType">The content type of the attachment.</param>
+    public void AddAttachment(
+        string filePath,
+        AttachmentType type = AttachmentType.Default,
+        string? contentType = null)
+    {
+        if (_options is not null)
+        {
+            _attachments.Add(
+                new Attachment(
+                    type,
+                    new FileAttachmentContent(filePath, _options.UseAsyncFileIO),
+                    Path.GetFileName(filePath),
+                    contentType));
+        }
+    }
 
     /// <summary>
     /// Creates a new Hint with one or more attachments.