+++ test/Sentry.Tests/ScopeTests.cs
@@ -98,7 +98,9 @@ public class ScopeTests
 
         var clone = sut.Clone();
 
-        Assert.Same(propagationContext, clone.PropagationContext);
+        Assert.NotSame(propagationContext, clone); // Sanity check that it really is a clone
+        Assert.Equal(propagationContext.TraceId, clone.PropagationContext.TraceId);
+        Assert.Equal(propagationContext.SpanId, clone.PropagationContext.SpanId);
     }
 
     [Fact]